+++ src/main/java/com/bit/backend/dtos/StockDto.java
@@ -3,13 +3,13 @@ package com.bit.backend.dtos;
 public class StockDto {
     private  Long id;
     private int stockItemID;
-    private int qty;
+    private Double qty;
     private String stockItemName;
 
     public StockDto() {
     }
 
-    public StockDto(Long id, int stockItemID, int qty, String stockItemName) {
+    public StockDto(Long id, int stockItemID, Double qty, String stockItemName) {
         this.id = id;
         this.stockItemID = stockItemID;
         this.qty = qty;
@@ -32,11 +32,11 @@ public class StockDto {
         this.stockItemID = stockItemID;
     }
 
-    public int getQty() {
+    public Double getQty() {
         return qty;
     }
 
-    public void setQty(int qty) {
+    public void setQty(Double qty) {
         this.qty = qty;
     }
